+++ src/Data/CheckoutPage.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\SchemaOrg\Data;
+
+class CheckoutPage extends WebPage
+{
+}