+++ CHANGELOG.md
@@ -1,7 +1,31 @@
-1.0.1 (unreleased)
+1.1.0
 =====
 
 *   (improvement) Initialize `ContactPoint::$contactType` and `PostalAddress::$contactType` with a default value.
+*   (feature) Added new data types (logically grouped):
+    - `Answer` 
+    - `Comment`
+    - `Question`
+    - `WebPage`
+    - `WebPageElement`
+    - `AboutPage`
+    - `CheckoutPage`
+    - `CollectionPage`
+    - `ContactPage`
+    - `FAQPage`
+    - `ItemPage`
+    - `ProfilePage`
+    - `RealEstateListing`
+    - `SearchResultsPage`
+    - `Intangible`
+    - `PriceSpecification`
+    - `Demand`
+    - `Offer`
+*   (improvement) Added `Product:$offers` and `Product::gtin*`.
+*   (feature) Added enum types:
+    - `BusinessFunction`
+    - `ItemAvailability`
+    - `PaymentMethod`
 
 
 1.0.0